fix(desktop): split observer store under the file-size ratchet
Main's accepted-event replay comments pushed observerRelayStore.ts
over 1000 lines. Move event ordering and the lifecycle watermark into
sibling modules so the merge stays under the ceiling.
